backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
python-dateutil>=2.8.2
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

